Adding Checklist formatting to step-by-step directions
Adding 'checklist' formatting to the step-by-step directions that
learners are required to follow in this episode. This requires changing
the callout code from ’challenge’ to ‘checklist’, and
rearranging/editing instructions to make the checklist match the exact
process in OpenRefine’s user interface (e.g. buttons). The first
sentence in the episode has moved to step 2. Please see
https://github.com/data-lessons/library-openrefine/issues/146 for
details about formatting changes.