Update 02-importing-data.md
Suggest  an additional question and two new objectives for the OpenRefine "Importing Data" episode. 

Issue number 68 opened a question about whether or not the lesson covers how to start and run OpenRefine. 
This is an issue that raises questions for many people since the software runs in the browser, and it is 
usually covered at the end of episode 1 or beginning of episode 2. Therefore, respectfully requesting that 
this should be included as a listed objective. The episode also covers basic functions of the  OR opening screen.