changes to objectives for the lesson to student outcomes
Edit as part of instructor checkout. With online materials where a student may return to refresh themselves at any point its good to make the key learning outcomes for the student rather than objectives. This will mean more to the student when they return and also reduce the number of objectives as objectives often end up as a long list of each thing covered, As a guide write 'what will the student be able to do after this lesson they couldn't before?'