Remove non-working link 13-looking-up-data.md
It appears that the cross function is no longer supplied by VIB.  Or at least I can't find it via either the VIB website, or the openrefine website.  If someone else knows where to find it, perhaps this can be added back with the correct link.  It sounds like a very valuable extension.