Update 04-faceting-and-filtering.md
Added " based on a unique text string, like a 'find' feature." to line 59 as per issue raised #137 https://github.com/LibraryCarpentry/lc-open-refine/issues/137. Added to this branch as per request from Talisha
"Hi Amanda,

Thanks for your email. Would you mind making your lesson contribution directly on the lesson branch so the maintainers might review and possibly use your suggestion in the lesson? Feel free to file it as a pull request directly on the repository or as an issue.

https://github.com/LibraryCarpentry/lc-open-refine

Best,
Talisha"