6. Working with columns and sorting
This is a suggested change to a lesson as part of checkout as a new instructor. I hope I'm doing it correctly!

Some of the sentences in "sorting data" are confusing to me.

"Once you have sorted the data, a new sort drop-down menu will be displayed"

I think it means that the sort option for that specific column now has the reverse and remove option? If yes, I suggest it gets removed altogether because the next line explains these new options.

The last sentence is also a touch confusing. Maybe wording such as this would fix it?

"To sort multiple columns, click on the drop-down menu for the second column you want to sort and choose 'sort'. In the new window, make sure that 'sort by this column alone' is not checked."